Autorização de usuários pelas abilities do token com o sanctum

+++ app/Http/Controllers/Api/SeriesController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\SeriesFormRequest;
 use App\Models\Series;
 use App\Repositories\SeriesRepository;
+use Illuminate\Contracts\Auth\Authenticatable;
 use Illuminate\Http\Request;
 
 class SeriesController extends Controller
@@ -45,9 +46,12 @@ class SeriesController extends Controller
         return $series;
     }
 
-    public function destroy(int $series)
+    public function destroy(int $series, Authenticatable $user)
     {
-        Series::destroy($series);
-        return response()->noContent();
+        if($user->tokenCan('series:delete')) {
+            Series::destroy($series);
+            return response()->noContent();
+        }
+        return response()->json(['message' => 'Unauthorized'], 401);
     }
 }
